feat: support dynamic work buffer formats in compute GSplat renderers (#8547)
* feat: support dynamic work buffer formats in compute GSplat renderers

Compute tiled renderers now dynamically support different work buffer
formats instead of hardcoding the compact format's bindings and decode
logic. Format-specific shader code is generated from GSplatFormat
metadata, with shader recreation on format changes at runtime.

Made-with: Cursor

* Address PR feedback: derive bindings, bind extra streams, add getOpacity(), repack compact format, getColor() returns vec3

- Derive formatStartBinding from fixed bindings array length instead of hardcoding
- Include extra streams (pcId, pcNodeIndex) in compute bind formats, declarations, and dispatch loops
- Add getOpacity() to work buffer format readers for early opacity culling before loading color
- Repack compact format: co-locate alpha with center in dataTransformA, move rotation to dataTransformB
- Change getColor() to return vec3 (RGB only) for both work buffer formats; opacity handled separately
- Remove conditional colorLoaded branch in packed format reader; getOpacity() unconditionally caches color
- Add GSPLAT_SEPARATE_OPACITY define for quad VS to call getOpacity() before getColor()
- Consolidate shader recreation to avoid redundant double-recreation per frame

Made-with: Cursor

* half precision

---------

+++ src/scene/shader-lib/wgsl/chunks/gsplat/compute-gsplat-tile-intersect.js
@@ -18,14 +18,14 @@ struct SplatTileEval {
     splatMax: vec2f,
 }
 
-fn computeRadiusFactor(opacity: f32) -> f32 {
-    return min(8.0, 2.0 * log(255.0 * opacity));
+fn computeRadiusFactor(opacity: half) -> f32 {
+    return min(8.0, 2.0 * log(255.0 * f32(opacity)));
 }
 
 fn computeSplatTileEval(
     screen: vec2f,
     coeffX: f32, coeffY: f32, coeffXY: f32,
-    opacity: f32,
+    opacity: half,
     viewportWidth: f32, viewportHeight: f32
 ) -> SplatTileEval {
     let K = 4.0 * coeffX * coeffY - coeffXY * coeffXY;
